requirements.txt
# Core dependencies for 3D reconstruction pipeline
numpy>=1.24.0
laspy>=2.5.0
lazrs>=0.6.0
open3d>=0.17.0
pillow>=10.0.0
opencv-python>=4.8.0
trimesh>=4.0.0
pyproj>=3.6.0
scipy>=1.11.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.66.0
pyyaml>=6.0.1
requests>=2.31.0
